Move more code out of `__init__` file
